Initial Files
Here are the files for my crypto bot. This is only a demo for now. It uses CoinGecko to get crypto prices, and set values for investment amounts. You can tinker with these settings in the config.json file. I will be using Binance API in the future when I finish with the foundation. Once that is done the project might be renamed.